workflow_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
multipart
kafka-python